Moved API Client from Services to Client namespace
